pyproject.toml
[project]
name = "# DB Querying Agent MCP Server"
version = "0.1.0"
description = "Database querying AI agent as a Model Context Protocol (MCP)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.0.2",
    "openai>=1.109.1",
    "psycopg2-binary>=2.9.10",
    "rich>=14.1.0",
]
